IMPALA-8700: Use int64_t mtime for HdfsScanNodeBase::AllocateScanRange()
HDFS defines the modification time as the milliseconds since Jan 1st,
1970. Since milliseconds wrap every 49 days in a 4 byte integer,
all locations storing modification time need to use an 8 byte integer.
The overload of HdfsScanNodeBase::AllocateScanRange() that is called
from Impala-lzo uses a 4 byte integer. When this wraps, it triggers
a DCHECK in ScanRange::Reset() that assumes mtime > 0.

This changes the HdfsScanNodeBase::AllocateScanRange() overload to
take an int64_t. The Impala-lzo side already supplies an int64_t.

Reviewed-on: http://gerrit.cloudera.org:8080/13712
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>

+++ be/src/exec/hdfs-scan-node-base.cc
@@ -634,7 +634,7 @@ ScanRange* HdfsScanNodeBase::AllocateScanRange(hdfsFS fs, const char* file,
 
 ScanRange* HdfsScanNodeBase::AllocateScanRange(hdfsFS fs, const char* file,
     int64_t len, int64_t offset, int64_t partition_id, int disk_id, bool try_cache,
-    bool expected_local, int mtime, bool is_erasure_coded,
+    bool expected_local, int64_t mtime, bool is_erasure_coded,
     const ScanRange* original_split) {
   return AllocateScanRange(fs, file, len, offset, partition_id, disk_id, expected_local,
       is_erasure_coded, mtime, BufferOpts(try_cache), original_split);
