+++ be/src/exec/hdfs-scan-node-base.h
@@ -278,7 +278,7 @@ class HdfsScanNodeBase : public ScanNode {
   /// Old API for compatibility with text scanners (e.g. LZO text scanner).
   io::ScanRange* AllocateScanRange(hdfsFS fs, const char* file, int64_t len,
       int64_t offset, int64_t partition_id, int disk_id, bool try_cache,
-      bool expected_local, int mtime,
+      bool expected_local, int64_t mtime,
       bool is_erasure_coded = false,
       const io::ScanRange* original_split = NULL);
